+++ CHANGES.rst
@@ -1,6 +1,10 @@
 2.6.1dev
 =====================
+<<<<<<< sso_query
 - Update search URL for query_solar_system_objects
+=======
+- Fixed search result sort order for the cases involving TESS sectors 99 and 100, etc. [#1558]
+>>>>>>> main
 
 2.6.0 (2026-04-16)
 =====================